refactor: 调整 AccountSwitcher 组件位置
